Making changes to time services drawer
